before merge with master

+++ src/cmds/kick.cpp
@@ -6,7 +6,7 @@
 /*   By: aandric <aandric@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/14 14:57:55 by mgolinva          #+#    #+#             */
-/*   Updated: 2023/03/06 16:32:58 by aandric          ###   ########.fr       */
+/*   Updated: 2023/03/06 16:37:05 by aandric          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,7 +64,7 @@ void	Server::Kick(User &user, Message &message)
     if (message._argsNb > 3)
     {
         _rplMsg = "";
-        for (int i = 3; i != message._argsNb; i++) // check if _argsNb one argument = 1 || one argument = 0
+        for (size_t i = 3; i != message._argsNb; i++) // check if _argsNb one argument = 1 || one argument = 0
             _rplMsg = _rplMsg + message._arguments[i]; // build reply message with users' arguments
         _rplMsg = user.getNickname() + " @ IRC_MAXANA KICK #" + channel_name + nickname;
         send(user.getSockfd(), _errMsg.c_str(), _errMsg.length(), 0);
